requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.4
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.2
django-debug-toolbar==3.2.1
django-grappelli==2.15.1
django-modeltranslation==0.16.1
django-silk==4.1.0
future==0.18.2
gprof2dot==2021.2.21
html5lib==1.1
idna==2.10
Jinja2==2.11.3
MarkupSafe==2.0.1
Pillow==8.2.0
pycodestyle==2.7.0
Pygments==2.8.1
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-gettext==4.0
pytz==2021.1
reportlab==3.5.67
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
webencodings==0.5.1

celery~=5.1.0
djangorestframework~=3.12.4
pathlib~=1.0.1